+++ vite-project/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "lint": "^0.8.19",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.9.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.27",
