+++ vite-project/src/App.css
@@ -1,5 +1,6 @@
 #root {
-  max-width: 1280px;
+  /* max-width: 1280px; */
+  width: 100%;;
   margin: 0 auto;
   padding: 2rem;
   text-align: center;
