+++ vite-project/src/App.tsx
@@ -1,47 +1,24 @@
 import React from 'react';
-// import { useState } from 'react';
-// import reactLogo from './assets/react.svg';
 import './App.css';
-import SearchInput from './components/search-input';
-import Tile from './components/tile/tile';
-import data from './assets/data';
+import { Routes, Route } from 'react-router-dom';
+import Home from './pages/Home';
+import About from './pages/About';
+import NotFound from './pages/NotFound';
+import Layout from './components/Layout/Layout';
+// import SearchInput from './components/search-input';
+// import Tile from './components/tile/tile';
+// import data from './assets/data';
 
 export default function App() {
-  // const [count, setCount] = useState(0);
-  const nodes = getNodes();
   return (
-    <div className="App">
-      <SearchInput />
-      <div className="tiles">
-        {nodes.map((node) => (
-          <Tile key={node.id} node={node} />
-        ))}
-      </div>
-      {/* <div>
-        <a href="https://vitejs.dev" target="_blank">
-          <img src="/vite.svg" className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://reactjs.org" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-        */}
-    </div>
+    <>
+      <Routes>
+        <Route path="/" element={<Layout />}>
+          <Route index element={<Home />} />
+          <Route path="about" element={<About />} />
+          <Route path="*" element={<NotFound />} />
+        </Route>
+      </Routes>
+    </>
   );
-}
-
-function getNodes() {
-  return data;
 }
