+++ vite-project/src/components/Layout/layout.module.scss
@@ -1,0 +1,9 @@
+.header {
+  font-size: 2rem;
+  background-color: #242424;
+  width: 100%;
+  justify-content: center;
+  padding-bottom: 40px;
+  display: flex;
+  justify-content: space-evenly;
+}
