+++ vite-project/src/index.css
@@ -13,12 +13,23 @@
   -moz-osx-font-smoothing: grayscale;
   -webkit-text-size-adjust: 100%;
 }
+/* 
+.header {
+  font-size: 2rem;
+  background-color: #242424;
+  width: 100%;
+  justify-content: center;
+  padding-bottom: 40px;
+  display: flex;
+  justify-content: space-evenly;
+} */
 
 a {
   font-weight: 500;
   color: #646cff;
   text-decoration: inherit;
 }
+
 a:hover {
   color: #535bf2;
 }
@@ -26,7 +37,7 @@ a:hover {
 body {
   margin: 0;
   display: flex;
-  place-items: center;
+  /* place-items: center; */
   min-width: 320px;
   min-height: 100vh;
 }
