+++ vite-project/src/pages/About.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default class About extends React.Component {
+  render() {
+    return <div className="App">About us page</div>;
+  }
+}
