+++ vite-project/src/pages/NotFound.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default class NotFound extends React.Component {
+  render() {
+    return <div className="App">Not FOUND</div>;
+  }
+}
